+++ module/Shared/src/Domain/DomainEventInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ExtendsSoftware\ExaPHPExample\Shared\Domain;
+
+interface DomainEventInterface {}